done login and file organising
