+++ lib/view/user_auth/signup_form_2.dart
@@ -1,9 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_countdown_timer/countdown.dart';
-import 'package:flutter_countdown_timer/countdown_controller.dart';
-import 'package:wmsm_flutter/view/custom/widgets/custom_elevatedbutton.dart';
-import 'package:wmsm_flutter/view/custom/widgets/custom_outlinedbutton.dart';
-import 'package:wmsm_flutter/view/shared/passcode_field.dart';
+import 'package:wmsm_flutter/view/user_auth/widgets/signup_form2_widget.dart';
 
 class SignUpForm2 extends StatefulWidget {
   const SignUpForm2({super.key});
@@ -16,6 +12,8 @@ class _SignUpForm2State extends State<SignUpForm2> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
+      //*to avoid content push upward when soft keyboard appears
+      resizeToAvoidBottomInset: false,
       body: Stack(children: [
         Container(
           decoration: BoxDecoration(
@@ -59,7 +57,7 @@ class _SignUpForm2State extends State<SignUpForm2> {
                           top: Radius.circular(30),
                         ),
                       ),
-                      child: const ContentClass(),
+                      child: const SignUpForm2Widget(),
                     ),
                   ),
                 ],
@@ -68,151 +66,6 @@ class _SignUpForm2State extends State<SignUpForm2> {
           ),
         ),
       ]),
-    );
-  }
-}
-
-class OTPTimer extends StatefulWidget {
-  const OTPTimer({super.key});
-
-  @override
-  State<OTPTimer> createState() => _OTPTimerState();
-}
-
-class _OTPTimerState extends State<OTPTimer> {
-  final int duration = 30;
-  late CountdownController countdownController;
-
-  @override
-  void initState() {
-    super.initState();
-    countdownController = CountdownController(
-      duration: Duration(seconds: duration),
-      onEnd: () {
-        print('Countdown Ended');
-      },
-    );
-    countdownController.start(); //*start the timer
-  }
-
-  @override
-  void dispose() {
-    countdownController.dispose();
-    super.dispose();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Countdown(
-      countdownController: countdownController,
-      builder: (BuildContext context, Duration remaining) {
-        return CustomOutlinedButton(
-            iconData: null,
-            context: context,
-            onPressed: () {
-              print('Pressed resend link');
-            },
-            text: 'Resend (${remaining.inSeconds})');
-      },
-    );
-  }
-}
-
-class ContentClass extends StatefulWidget {
-  const ContentClass({super.key});
-
-  @override
-  State<ContentClass> createState() => _ContentClassState();
-}
-
-class _ContentClassState extends State<ContentClass> {
-  final TextEditingController _passcode = TextEditingController();
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: <Widget>[
-        Text(
-          'Sign Up',
-          style: Theme.of(context).textTheme.displayLarge,
-        ),
-        Text(
-          'Signing up is easy. Just fill up the details and you are set to go',
-          style: Theme.of(context).textTheme.bodyLarge,
-        ),
-        const SizedBox(
-          height: 70,
-        ),
-        Text(
-          'Phone number'.toUpperCase(),
-          style: Theme.of(context).textTheme.bodyLarge,
-        ),
-        const SizedBox(
-          height: 10,
-        ),
-        //*Phone number row
-        Row(
-          children: [
-            SizedBox(
-              width: 30,
-              height: 30,
-              child: Image.network(
-                'https://flagsapi.com/MY/flat/64.png',
-                errorBuilder: (BuildContext context, Object exception,
-                    StackTrace? stackTrace) {
-                  return Image.asset("assets/images/64.png");
-                },
-              ),
-            ),
-            Padding(
-              padding: const EdgeInsets.only(left: 10.0, right: 20.0),
-              child: Text(
-                '+6011-12345678', //TODO: Get phone number from user
-                style: Theme.of(context).textTheme.bodyLarge,
-              ),
-            ),
-            const OTPTimer(),
-          ],
-        ),
-        SizedBox(
-          height: MediaQuery.of(context).size.height * 0.05,
-        ),
-        //*OTP boxes
-        Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            Text(
-              'PASSCODE',
-              style: Theme.of(context).textTheme.bodyLarge,
-            ),
-            const SizedBox(
-              height: 10,
-            ),
-            passCodeField(
-              passcodeController: _passcode,
-            ),
-          ],
-        ),
-        SizedBox(
-          height: MediaQuery.of(context).size.height * 0.05,
-        ),
-        Row(
-          children: [
-            Expanded(
-              child: CustomElevatedButton(
-                onPressed: () {
-                  print('Pressed continue');
-                  Navigator.pushNamed(context, '/f3');
-                },
-                child: Text(
-                  'Continue'.toUpperCase(),
-                ),
-              ),
-            ),
-          ],
-        ),
-      ],
     );
   }
 }
