+++ lib/view/user_auth/signup_form_4.dart
@@ -1,8 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:wmsm_flutter/view/custom/widgets/custom_elevatedbutton.dart';
-import 'package:wmsm_flutter/view/shared/passcode_field.dart';
-
-import '../custom/widgets/custom_outlinedbutton.dart';
+import 'package:wmsm_flutter/view/user_auth/widgets/signup_form4_widget.dart';
 
 class SignUpForm4 extends StatelessWidget {
   const SignUpForm4({super.key});
@@ -10,6 +7,7 @@ class SignUpForm4 extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
+      resizeToAvoidBottomInset: false,
       body: Stack(children: [
         Container(
           decoration: BoxDecoration(
@@ -55,7 +53,7 @@ class SignUpForm4 extends StatelessWidget {
                           top: Radius.circular(30),
                         ),
                       ),
-                      child: const WidgetSignUp4(),
+                      child: const SignUpForm4Widget(),
                     ),
                   ),
                 ],
@@ -64,78 +62,6 @@ class SignUpForm4 extends StatelessWidget {
           ),
         ),
       ]),
-    );
-  }
-}
-
-class WidgetSignUp4 extends StatefulWidget {
-  const WidgetSignUp4({super.key});
-
-  @override
-  State<WidgetSignUp4> createState() => _WidgetSignUp4State();
-}
-
-class _WidgetSignUp4State extends State<WidgetSignUp4> {
-  final TextEditingController _passcode = TextEditingController();
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: [
-        const Text(
-          'A validation code has been sent to the mailbox of eunicelim1520@gmail.com',
-          textAlign: TextAlign.justify,
-        ),
-        const SizedBox(
-          height: 40,
-        ),
-        Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            Text(
-              'Verification Code',
-              style: Theme.of(context).textTheme.bodyLarge,
-            ),
-            const SizedBox(
-              height: 30,
-            ),
-            passCodeField(
-              passcodeController: _passcode,
-            ),
-          ],
-        ),
-        const SizedBox(
-          height: 40,
-        ),
-        Row(
-          children: [
-            Expanded(
-              child: CustomOutlinedButton(
-                iconData: null,
-                context: context,
-                onPressed: () => print('next page'),
-                text: 'RESEND LINK',
-              ),
-            ),
-          ],
-        ),
-        const SizedBox(
-          height: 10,
-        ),
-        Row(
-          children: [
-            Expanded(
-              child: CustomElevatedButton(
-                  onPressed: () {
-                    print(_passcode.text.trim());
-                    Navigator.pushNamed(context, '/f5');
-                  },
-                  child: const Text('CONTINUE')),
-            ),
-          ],
-        )
-      ],
     );
   }
 }
